+++ src/genome/__init__.py
@@ -2,6 +2,8 @@
 
 from importlib.metadata import PackageNotFoundError, version
 
+from genome.genome import Genome
+from genome.region import Region
 from genome.seq import DNA, RNA, Protein
 
 try:
@@ -9,4 +11,4 @@ try:
 except PackageNotFoundError:
     __version__ = "0.0.0+unknown"
 
-__all__ = ["DNA", "RNA", "Protein", "__version__"]
+__all__ = ["DNA", "RNA", "Genome", "Protein", "Region", "__version__"]